Better flex handling - add support for wrinkle maps - add support for "side" uint8 variable for stereo flex balances - create "partner flexes" all the time, not just when making drivers

+++ library/models/mdl/structs/flex.py
@@ -251,6 +251,6 @@ class VertAnimWrinkleV49(VertAnimV49):
             ('side', np.uint8, (1,)),
             ('vertex_delta', np.float16, (3,)),
             ('normal_delta', np.float16, (3,)),
-            ('wrinkle_delta', np.float16, (1,)),
+            ('wrinkle_delta', np.int16, (1,)),
         ]
     )
